% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abduce.R
\name{abduce}
\alias{abduce}
\title{Abduce}
\usage{
abduce(bottom_clauses, hypothesis, head_clauses = NULL)
}
\arguments{
\item{bottom_clauses}{list produced by the \code{\link{get_bottom_clause}} function.}

\item{hypothesis}{character vector where each element relates the different observations produced by the get_observation functions.}

\item{head_clauses}{character vector with new head clauses that override the ones present in the bottom clause list.}
}
\value{
a data frame where each row contains a pair of OTU, the effect on the abundance and the compression supporting the effect.
}
\description{
Uses PyGol abduction procedure to get the compression values of the different effects
}
\examples{
abduce(bottom_clause, hypothesis)

}
