% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_OTU_table}
\alias{check_OTU_table}
\title{Check OTU table
Checks the presence of NA and non-numeric cells}
\usage{
check_OTU_table(otu_tab)
}
\arguments{
\item{otu_tab}{data frame with the OTUs counts with samples on columns and OTUs in rows.}
}
\value{
revised data frame
}
\description{
Check OTU table
Checks the presence of NA and non-numeric cells
}
\examples{
check_OTU_table(otu_tab)
}
