% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_interactions.R
\name{classify_interactions}
\alias{classify_interactions}
\title{Classify interactions
Automatic classification of significant effects to obtain the types of interacions}
\usage{
classify_interactions(effect_table)
}
\arguments{
\item{effect_table}{a data frame, containing the significant effects where each row contains a pair of OTU, the effect on the abundance and the I value supporting the effect.}
}
\value{
a data frame, similar to effect_table, where effect_up and effect_down has been changed by the types of interactions proposed by Derocles et al. (2018)
}
\description{
Classify interactions
Automatic classification of significant effects to obtain the types of interacions
}
\examples{
classify_interactions(effect_table)
}
