% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compositional_observations.R
\name{do_chi_abundance}
\alias{do_chi_abundance}
\title{Perform Chi Square test
Generate a logical clause depicting the compositional abundance change between two samples.}
\usage{
do_chi_abundance(comparison, otu_abundance, depth, exclusion = FALSE)
}
\arguments{
\item{comparison}{numeric vector with the position, in the OTU table , of the samples to compare}

\item{otu_abundance}{list joining all the OTU information. Obtained using the join_abundances function.}

\item{depth}{numeric vector with the sequencing depth}

\item{exclusion}{logical indicating if exclusion cases have to be considered for generating the logic clauses. Default is FALSE.}
}
\value{
character with the abundance clause describing the compositional abundance change between samples.
}
\description{
Perform Chi Square test
Generate a logical clause depicting the compositional abundance change between two samples.
}
\examples{
do_chi_abundance(comparison, otu_abundance, depth)
}
