% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatOutput}
\alias{formatOutput}
\title{Format abduction output
Gives data frame format to the raw abduction output produced by PyGol}
\usage{
formatOutput(abduced_raw)
}
\arguments{
\item{abduced_raw}{data frame containing the raw output of PyGol abduction and the compression values.}
}
\value{
data frame where each row contains a pair of OTU, the effect on the abundance and the compression supporting the effect.
}
\description{
Format abduction output
Gives data frame format to the raw abduction output produced by PyGol
}
\examples{
formatOutput(abduced_raw)
}
