% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bottom_clause.R
\name{get_bottom_clause}
\alias{get_bottom_clause}
\title{Generate bottom clause
Uses PyGol to generate the bottom clauses from the head and body observations of the hypothesis of interaction.}
\usage{
get_bottom_clause(otu_data, head_clauses, body_clauses, search.depth = 2)
}
\arguments{
\item{otu_data}{list joining all the OTU information. Obtained using the join_abundances function.}

\item{head_clauses}{character vector with the clauses in the head position of the hypothesis of interaction.}

\item{body_clauses}{character vector  with the clauses in the body position of the hypothesis of interaction.}

\item{search.depth}{numeric selecting how deep has to search PyGol to generate the abduction clauses.}
}
\value{
list including a python dictionary with the bottom clauses together with all the information required for the abduction.
}
\description{
Generate bottom clause
Uses PyGol to generate the bottom clauses from the head and body observations of the hypothesis of interaction.
}
\examples{
get_bottom_clause(otu_data, head_clauses, body_clauses)
}
