% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_comparsions}
\alias{get_comparsions}
\title{Get comparisons
Produces a list with all the possible comparisons between samples in both directions.}
\usage{
get_comparsions(n_samples)
}
\arguments{
\item{n_samples}{numeric vector with the position of tha samples to compare.}
}
\value{
a list with the pairs of samples to compare.
}
\description{
Get comparisons
Produces a list with all the possible comparisons between samples in both directions.
}
\examples{
get_comparsions(n_samples)
}
