% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_compositional_observations.R
\name{get_compositional_observations}
\alias{get_compositional_observations}
\title{Get compositional observation
Generate logical clauses depicting the compositional abundance change between two samples.}
\usage{
get_compositional_observations(
  otu_abundance,
  comparisons,
  depth,
  exclusion = FALSE
)
}
\arguments{
\item{otu_abundance}{list joining all the OTU information. Obtained using the join_abundances function.}

\item{comparisons}{list with the position of the pairs of samples to compare.}

\item{depth}{numeric vector with the sequencing depth}

\item{exclusion}{logical indicating if exclusion cases have to be considered for generating the logic clauses. Default is FALSE.}
}
\value{
character vector with the abundance clauses describing the compositional abundance change between samples.
}
\description{
Get compositional observation
Generate logical clauses depicting the compositional abundance change between two samples.
}
\examples{
get_compositional_observations(otu_abundance, comparisons, depth)
}
