% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_I_values.R
\name{get_I_values}
\alias{get_I_values}
\title{Compute the I stastitc
Obtain the I statistic from the compression values result of abduction}
\usage{
get_I_values(abduced)
}
\arguments{
\item{abduced}{data frame where each row contains a pair of OTU, the effect on the abundance and the compression supporting the effect. Result of abduce function.}
}
\value{
data frame where each row contains a pair of OTU, the effect on the abundance and the I statistic supporting the effect.
}
\description{
Compute the I stastitc
Obtain the I statistic from the compression values result of abduction
}
\examples{
get_I_values(abduced)
}
