% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence.R
\name{get_presence}
\alias{get_presence}
\title{Get presence clauses
Generate logical clauses depicting the presence of each OTU in each sample}
\usage{
get_presence(otu_data)
}
\arguments{
\item{otu_data}{list joining all the OTU information. Obtained using the join_abundances function.}
}
\value{
character vector with the presence clauses describing the presence of the OTUs in the samples.
}
\description{
Get presence clauses
Generate logical clauses depicting the presence of each OTU in each sample
}
\examples{
get_presence(otu_data)
}
