% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinte.R
\name{infinte}
\alias{infinte}
\title{InfIntE
InfIntE (INFerence of INTeractions using Explainable machine learning). This function contains the full InfIntE pipeline. It uses compositional and/or absolute OTU data to infer ecological interactions using a logical hypothesis of interaction.
The InfIne inference process parses the OTU data into logic clauses and performs the abduction of interactions using PyGol. Interactions are selected using pulsar.}
\usage{
infinte(
  otu_tb,
  thresh = 0.01,
  exclusion = FALSE,
  nperms = 50,
  search.depth = 2,
  hypothesis = NULL,
  depth = NULL,
  absolute_abundance = NULL,
  ncores = 1
)
}
\arguments{
\item{otu_tb}{data frame with the OTUs counts with samples on columns and OTUs in rows.}

\item{thresh}{numeric threshold for StARS model selection in pulsar, default is 0.01}

\item{exclusion}{logical indicating if exclusion cases have to be considered for generating the logic clauses. Default is FALSE.}

\item{nperms}{numeric number of subsamples in pulsar, default is 50}

\item{search.depth}{numeric search depth when producing constructing the bottom clauses, default is 2}

\item{hypothesis}{character vector where each element relates the different observations produced by the get_observation functions.}

\item{depth}{numeric vector containing an alternative sequence depth in case the compositional OTU table is a subset of one with more OTUs}

\item{absolute_abundance}{data frame with the absolute counts with samples on columns and OTUs in rows.}

\item{ncores}{number of cores used by pulsar}
}
\value{
list with the infinite output. Contains a data frame with the interactions inferred by InfIntE, the pulsar output and the abduction output.
}
\description{
InfIntE
InfIntE (INFerence of INTeractions using Explainable machine learning). This function contains the full InfIntE pipeline. It uses compositional and/or absolute OTU data to infer ecological interactions using a logical hypothesis of interaction.
The InfIne inference process parses the OTU data into logic clauses and performs the abduction of interactions using PyGol. Interactions are selected using pulsar.
}
\examples{
infinte(otu_tb, hypothesis, thresh = 0.01, exclusion = FALSE, nperms = 50, search.depth = 2, depth = NULL, absolute_abundance = NULL)
}
