% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_abundances.R
\name{join_abundances}
\alias{join_abundances}
\title{Join abundances
Joins in a single data frame the compositional and absolute OTU counts and returns a list with the new data frame and all the information of samples and OTUs}
\usage{
join_abundances(otu_tb, absolute_abundance, depth = NULL)
}
\arguments{
\item{otu_tb}{data frame with the OTUs counts with samples on columns and OTUs in rows.}

\item{absolute_abundance}{data frame with the absolute counts with samples on columns and OTUs in rows.}

\item{depth}{numeric vector alternative sequencing depth in case OTU table is a subset.}
}
\value{
list joining all the OTU information including an OTU table with absolute and/or compositional sequence counts.
}
\description{
Join abundances
Joins in a single data frame the compositional and absolute OTU counts and returns a list with the new data frame and all the information of samples and OTUs
}
\examples{
join_abundances(otu_tb, absolute_abundance, depth = NULL)
}
