% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_PyGol}
\alias{load_PyGol}
\title{Load PyGol
Loads PyGol function for abduction using reticulate.}
\usage{
load_PyGol()
}
\description{
Load PyGol
Loads PyGol function for abduction using reticulate.
}
\examples{
load_PyGol()
}
