% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pulsar_infinte.R
\name{pulsar_infinte}
\alias{pulsar_infinte}
\title{InfInte function for pulsar
Function to be used by pulsar to perform the abduction from a subset of the OTU table and obtain a list of adjacency matrix along the lambda path}
\usage{
pulsar_infinte(
  sub_otu_tb,
  lambda,
  bottom_clauses,
  hypothesis,
  exclusion,
  otu_data
)
}
\arguments{
\item{sub_otu_tb}{subset of the OTU table provided by pulsar}

\item{lambda}{numeric vector containing the lambda path obtained with the \code{\link{getLamPath}} function.}

\item{bottom_clauses}{list including a python dictionary with the bottom clauses together with all the information required for the abduction.}

\item{hypothesis}{character vector where each element relates the different observations produced by the get_observation functions.}

\item{exclusion}{logical indicating if exclusion cases have to be considered for generating the logic clauses. Default is FALSE.}

\item{otu_data}{list joining all the OTU information. Obtained using the join_abundances function.}
}
\value{
a list of p∗p adjacency matrices, one for each value of lambda.
}
\description{
InfInte function for pulsar
Function to be used by pulsar to perform the abduction from a subset of the OTU table and obtain a list of adjacency matrix along the lambda path
}
\examples{
pulsar_infinte(sub_otu_tb, lambda, bottom_clauses, hypothesis, exclusion, otu_data)
}
