% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_names}
\alias{return_names}
\title{Return names
Returns de original names to the OTUs}
\usage{
return_names(infered, otu_names)
}
\arguments{
\item{otu_names}{named character vector with the original otu names in the same order as provided in the original input.}

\item{abduced}{a data frame, containing the significant effects where each row contains a pair of OTU, the effect on the abundance or interaction and the I value supporting the effect.}
}
\value{
the same input data frame substituting the executing names by the original ones.
}
\description{
Return names
Returns de original names to the OTUs
}
\examples{
return_names(infered, otu_names)
}
